Added CommentRepository.GetComments()

+++ BlogEngine/Models/Entities/Comment.cs
@@ -2,9 +2,9 @@
 {
     public class Comment : BaseEntity
     {
-        public int ID { get; set; }
+        public long ID { get; set; }
         public string Title { get; set; }
         public string Body { get; set; }
-        public int PostID { get; set; }
+        public long PostID { get; set; }
     }
 }
